@@ -46,50 +46,50 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "14.2.5",
-        "@angular/common": "14.2.5",
-        "@angular/compiler": "14.2.5",
-        "@angular/core": "14.2.5",
-        "@angular/forms": "14.2.5",
-        "@angular/platform-browser": "14.2.5",
-        "@angular/platform-browser-dynamic": "14.2.5",
-        "@angular/router": "14.2.5",
-        "@nestjs/common": "9.1.4",
-        "@nestjs/core": "9.1.4",
-        "@nestjs/platform-express": "9.1.4",
-        "@nrwl/angular": "14.8.3",
+        "@angular/animations": "15.0.0",
+        "@angular/common": "15.0.0",
+        "@angular/compiler": "15.0.0",
+        "@angular/core": "15.0.0",
+        "@angular/forms": "15.0.0",
+        "@angular/platform-browser": "15.0.0",
+        "@angular/platform-browser-dynamic": "15.0.0",
+        "@angular/router": "15.0.0",
+        "@nestjs/common": "9.2.0",
+        "@nestjs/core": "9.2.0",
+        "@nestjs/platform-express": "9.2.0",
+        "@nrwl/angular": "15.2.1",
         "reflect-metadata": "^0.1.13",
         "rxjs": "7.5.7",
-        "tslib": "^2.4.0",
+        "tslib": "^2.4.1",
         "zone.js": "0.11.8"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "14.2.5",
-        "@angular-eslint/eslint-plugin": "14.0.4",
-        "@angular-eslint/eslint-plugin-template": "14.0.4",
-        "@angular-eslint/template-parser": "14.0.4",
-        "@angular/cli": "14.2.5",
-        "@angular/compiler-cli": "14.2.5",
-        "@angular/language-service": "14.2.5",
+        "@angular-devkit/build-angular": "15.0.0",
+        "@angular-eslint/eslint-plugin": "14.3.0",
+        "@angular-eslint/eslint-plugin-template": "14.3.0",
+        "@angular-eslint/template-parser": "14.3.0",
+        "@angular/cli": "15.0.0",
+        "@angular/compiler-cli": "15.0.0",
+        "@angular/language-service": "15.0.0",
         "@cypress/webpack-dev-server": "2.3.0",
         "@nestjs/schematics": "9.0.3",
         "@nestjs/testing": "9.1.4",
-        "@nrwl/cli": "14.8.3",
-        "@nrwl/cypress": "14.8.3",
-        "@nrwl/eslint-plugin-nx": "14.8.3",
-        "@nrwl/jest": "14.8.3",
-        "@nrwl/linter": "14.8.3",
-        "@nrwl/nest": "14.8.3",
-        "@nrwl/node": "14.8.3",
-        "@nrwl/workspace": "14.8.3",
+        "@nrwl/cli": "15.2.1",
+        "@nrwl/cypress": "15.2.1",
+        "@nrwl/eslint-plugin-nx": "15.2.1",
+        "@nrwl/jest": "15.2.1",
+        "@nrwl/linter": "15.2.1",
+        "@nrwl/nest": "15.2.1",
+        "@nrwl/node": "15.2.1",
+        "@nrwl/workspace": "15.2.1",
         "@types/jest": "28.1.8",
         "@types/node": "18.7.21",
         "@types/supertest": "^2.0.12",
-        "@typescript-eslint/eslint-plugin": "5.36.1",
-        "@typescript-eslint/parser": "5.36.1",
-        "cypress": "10.10.0",
+        "@typescript-eslint/eslint-plugin": "5.43.0",
+        "@typescript-eslint/parser": "5.43.0",
+        "cypress": "11.1.0",
         "dotenv": "16.0.3",
-        "eslint": "8.24.0",
+        "eslint": "8.28.0",
         "eslint-config-prettier": "8.5.0",
         "eslint-plugin-cypress": "2.12.1",
         "html-webpack-plugin": "5.5.0",
@@ -97,9 +97,9 @@
         "jest-environment-jsdom": "28.1.1",
         "jest-extended": "^3.1.0",
         "jest-preset-angular": "12.2.2",
-        "nx": "14.8.3",
+        "nx": "15.2.1",
         "prettier": "2.7.1",
-        "supertest": "^6.3.0",
+        "supertest": "^6.3.1",
         "ts-jest": "28.0.8",
         "ts-node": "10.9.1",
         "typescript": "4.8.4"
@@ -107,5 +107,13 @@
     "engines": {
         "node": "16.17.0",
         "npm": "8.15.0"
+    },
+    "overrides": {
+        "jest-preset-angular": {
+            "@angular-devkit/build-angular": "15.0.0",
+            "@angular/compiler-cli": "15.0.0",
+            "@angular/core": "15.0.0",
+            "@angular/platform-browser-dynamic": "15.0.0"
+        }
     }
 }
