@@ -1,3 +1,4 @@
+import { mount } from 'cypress/angular';
 /// <reference types="cypress" />
 // ***********************************************
 // This example commands.ts shows you how to
@@ -10,10 +11,14 @@
 // ***********************************************
 
 // eslint-disable-next-line @typescript-eslint/no-namespace
-declare namespace Cypress {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    interface Chainable<Subject> {
-        login(email: string, password: string): void;
+declare global {
+    // eslint-disable-next-line @typescript-eslint/no-namespace
+    namespace Cypress {
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        interface Chainable<Subject> {
+            login(email: string, password: string): void;
+            mount: typeof mount;
+        }
     }
 }
 //
@@ -32,3 +37,5 @@ Cypress.Commands.add('login', (email, password) => {
 //
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+Cypress.Commands.add('mount', mount);
